04_sovereign_knowledge_rag/requirements.txt
chromadb
ollama
google-generativeai
langchain
pypdf
langchain-text-splitters
